[Summary] move summary related logic into a separate library
