+++ clang/include/clang/Summary/SummaryAttribute.h
@@ -1,5 +1,5 @@
-#ifndef LLVM_CLANG_SEMA_SEMASUMMARYATTRIBUTE_H
-#define LLVM_CLANG_SEMA_SEMASUMMARYATTRIBUTE_H
+#ifndef LLVM_CLANG_SUMMARY_SUMMARYATTRIBUTE_H
+#define LLVM_CLANG_SUMMARY_SUMMARYATTRIBUTE_H
 
 #include "clang/AST/Decl.h"
 #include "clang/ASTMatchers/ASTMatchFinder.h"
@@ -51,4 +51,4 @@ public:
 };
 } // namespace clang
 
-#endif // LLVM_CLANG_SEMA_SEMASUMMARYATTRIBUTEH
+#endif // LLVM_CLANG_SUMMARY_SUMMARYATTRIBUTEH
