+++ clang/lib/Sema/SemaDecl.cpp
@@ -57,8 +57,8 @@
 #include "clang/Sema/SemaSYCL.h"
 #include "clang/Sema/SemaSwift.h"
 #include "clang/Sema/SemaWasm.h"
-#include "clang/Sema/SummaryContext.h"
 #include "clang/Sema/Template.h"
+#include "clang/Summary/SummaryContext.h"
 #include "llvm/ADT/STLForwardCompat.h"
 #include "llvm/ADT/SmallPtrSet.h"
 #include "llvm/ADT/SmallString.h"
