+++ clang/lib/Summary/CMakeLists.txt
@@ -1,0 +1,14 @@
+set(LLVM_LINK_COMPONENTS
+  Core
+  Support
+  )
+
+add_clang_library(clangSummary
+  SummaryAttribute.cpp
+  SummaryConsumer.cpp
+  SummaryContext.cpp
+
+  LINK_LIBS
+  clangAST
+  clangSupport
+  )
